responsive layout fully operative within the app after login
