+++ app/javascript/controllers/dropdown_controller.js
@@ -13,3 +13,4 @@ const dropDown = () => {
 }
 
 dropDown();
+
