+++ app/views/layouts/application.html.erb
@@ -20,8 +20,12 @@
       <%= render 'shared/navbar_landing' %>
     <% end %>
     <%= yield %>
-    <%= javascript_include_tag 'application' %>
 
-    <%= javascript_pack_tag 'application' %>
+    <%= javascript_include_tag 'application' %>
+    <% if user_signed_in? %>
+      <%= javascript_pack_tag 'application' %>
+    <% else %>
+      <%= javascript_pack_tag 'dropdown_login' %>
+    <% end %>
   </body>
 </html>
