+++ app/views/shared/_navbar_landing.html.erb
@@ -1,4 +1,4 @@
-<nav>
+<nav id="navbar_landing">
 
   <div class="navbar-mobily">
