+++ app/views/shared/_navbar_main.html.erb
@@ -1,4 +1,4 @@
-<nav>
+<nav id="navbar_main">
   <div class="navbar-mobily">
       <ul class="navbar-links">
         <li class="navbar-logo">
